[chore] Implement feedback

+++ server/src/controller/Product.js
@@ -3,16 +3,16 @@ import ProductModel from '../models/Product';
 class Product {
   static async getAllProducts(req, res) {
     const products = await ProductModel.findAll();
-    return res.status(200).json(products);
+    return res.status(200).json({ status: 'success', message: 'prodcuts fetched successfully', products });
   }
 
   static async getOneProduct(req, res) {
     const { id } = req.params;
     const product = await ProductModel.findOne(id);
     if (!product.value) {
-      return res.status(404).json({ message: 'product not found' });
+      return res.status(404).json({ status: 'error', message: 'product not found' });
     }
-    return res.send(product);
+    return res.status(200).json({ status: 'success', message: 'product fetched succesfully', product });
   }
 
   static async createProduct(req, res) {
@@ -21,18 +21,24 @@ class Product {
       res.status(400).json(newProduct);
       return;
     }
-    res.status(201).send({
-      success: true,
-      message: 'product created',
+    res.status(201).json({
+      status: 'success',
+      message: 'product created successfully',
       newProduct,
     });
   }
 
   static async deleteOneProduct(req, res) {
     const { id } = req.params;
+    const product = await ProductModel.findOne(id);
+    if (!product.value) {
+      res.status(400).json({ status: 'error', message: 'product not found' });
+      return;
+    }
     await ProductModel.deleteOne(id);
     res.status(200).json({
-      message: 'deleted',
+      status: 'success',
+      message: 'product deleted successfully',
     });
   }
 
@@ -41,13 +47,11 @@ class Product {
     const updatedProduct = await ProductModel.update(id, req.body);
     if (updatedProduct.errors.length !== 0) {
       if (updatedProduct.errors[0] === 'Product does not exist') {
-        res.status(404).send(updatedProduct);
+        res.status(404).json(updatedProduct);
         return;
       }
-      res.status(400).send(updatedProduct);
-      return;
     }
-    res.status(200).send(updatedProduct);
+    res.status(200).json({ status: 'success', message: 'product updated', updatedProduct });
   }
 }
 export default Product;
